upgrading micronaut and other dependencies

+++ non-blocking-pty/src/main/java/com/github/moaxcp/pty/Status.java
@@ -7,7 +7,7 @@ import lombok.Value;
 @Value
 @Builder
 public class Status {
-  int pid;
+  long pid;
   WinSize winSize;
   Integer result;
   boolean eventLoopRunning;
